rt - Update RoomSlots tests and use correct API endpoint
